Enforce exact academic data contracts end to end (#852)
* fix(server): enforce academic schema contracts

* fix(server): make academic response contracts exact

* ci: approve intentional GraphQL schema breaks

* fix(ci): prepare GraphQL check without database

* fix(ci): generate Prisma client for GraphQL check

* test(ci): fail closed on unreadable GraphQL base

* fix(mcp): enforce mode-specific output contracts

* fix(schema): reject nested MCP output drift

* fix(api): present exact schedule and metadata contracts

* fix(api): enforce subscription and homework contracts

+++ src/lib/graphql/schema.ts
@@ -142,7 +142,7 @@ export const graphqlTypeDefs = /* GraphQL */ `
     code: String!
     credits: Float
     period: Int
-    periodsPerWeek: Int
+    periodsPerWeek: Float
     timesPerWeek: Int
     stdCount: Int
     limitCount: Int
